entities are being integrated to pygame
