+++ scripts/constants_and_sources.py
@@ -2,6 +2,7 @@ import pygame
 display_width, display_height = (1440, 800)
 
 directions = 'left right up down'.split()
+opposites = ['right', 'left', 'dowm', 'up']
 
 heretic_images = {i: pygame.image.load(f'../images/heretic_sprite_{i}.png') for i in directions}
